LayoutTests/platform/ios-simulator-wk2/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/selection-not-application-expected.txt


FAIL text field selection for the input hidden Type error
FAIL text field selection for the input email assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input datetime-local Type error
FAIL text field selection for the input date Type error
FAIL text field selection for the input month Type error
FAIL text field selection for the input week Type error
FAIL text field selection for the input time Type error
FAIL text field selection for the input number assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input range Type error
FAIL text field selection for the input color assert_equals: expected (object) null but got (number) 0
FAIL text field selection for the input checkbox Type error
FAIL text field selection for the input radio Type error
FAIL text field selection for the input file Type error
FAIL text field selection for the input submit Type error
FAIL text field selection for the input image Type error
FAIL text field selection for the input reset Type error
FAIL text field selection for the input button Type error
PASS text field selection for the input text 
PASS text field selection for the input search 
PASS text field selection for the input tel 
PASS text field selection for the input url 
PASS text field selection for the input password 
